% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-foreground.R
\name{count_foreground_lineage}
\alias{count_foreground_lineage}
\title{Count independent foreground lineages}
\usage{
count_foreground_lineage(tree, trait)
}
\arguments{
\item{tree}{A \code{phylo} tree with unique tips.}

\item{trait}{A data frame containing \code{species} and one or more binary traits.}
}
\value{
A named list of independent foreground-lineage counts.
}
\description{
Count independent foreground lineages
}
\examples{
tree <- ape::read.tree(text="((A:1,B:1):1,C:1);")
trait <- data.frame(species=c("A", "B", "C"), foreground=c(1, 1, 0))
count_foreground_lineage(tree, trait)
}
