% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{fill_node_labels}
\alias{fill_node_labels}
\title{Fill missing internal-node labels}
\usage{
fill_node_labels(phy, verbose = FALSE)
}
\arguments{
\item{phy}{A `phylo` tree.}

\item{verbose}{Whether to emit a progress message.}
}
\value{
A `phylo` tree with unique generated labels for missing nodes.
}
\description{
Fill missing internal-node labels
}
