% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{get_ancestor_num}
\alias{get_ancestor_num}
\title{Get ancestor node numbers}
\usage{
get_ancestor_num(phy, node_num)
}
\arguments{
\item{phy}{A `phylo` tree.}

\item{node_num}{A single integer node number.}
}
\value{
Ancestor node numbers ordered from parent to root.
}
\description{
Get ancestor node numbers
}
