% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfpcm.R
\name{get_bootstrap_table}
\alias{get_bootstrap_table}
\title{Build a bootstrap-support table}
\usage{
get_bootstrap_table(pcm_out, bootstrap_result, mode = "l1ou")
}
\arguments{
\item{pcm_out}{An l1ou result object.}

\item{bootstrap_result}{An object containing `detection.rate`.}

\item{mode}{Processing mode; currently only `"l1ou"`.}
}
\value{
A node-name and bootstrap-support data frame.
}
\description{
Build a bootstrap-support table
}
