% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-collapse.R
\name{get_deepest_node_num}
\alias{get_deepest_node_num}
\title{Select the deepest candidate node}
\usage{
get_deepest_node_num(tree, node_nums)
}
\arguments{
\item{tree}{A \code{phylo} tree.}

\item{node_nums}{Candidate integer node numbers.}
}
\value{
A single integer node number or \code{NA}.
}
\description{
Select the deepest candidate node
}
