% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-traversal.R
\name{get_descendent_num}
\alias{get_descendent_num}
\title{Get descendant node numbers}
\usage{
get_descendent_num(phy, node_num, leaf_only = FALSE)
}
\arguments{
\item{phy}{A \code{phylo} tree.}

\item{node_num}{One or more integer node numbers.}

\item{leaf_only}{Whether to return only tip nodes.}
}
\value{
A sorted integer vector of descendants.
}
\description{
Get descendant node numbers
}
