% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{get_rooted_newick}
\alias{get_rooted_newick}
\title{Root a tree at a scored MAD edge}
\usage{
get_rooted_newick(t, madr, rho)
}
\arguments{
\item{t}{A `phylo` tree.}

\item{madr}{A single edge index.}

\item{rho}{Root position proportions, one per edge.}
}
\value{
A list containing Newick text, the rooted tree, and clock CV.
}
\description{
Root a tree at a scored MAD edge
}
