% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{get_single_branch_tree}
\alias{get_single_branch_tree}
\title{Construct a one-tip tree}
\usage{
get_single_branch_tree(name, dist)
}
\arguments{
\item{name}{A single non-empty tip label.}

\item{dist}{A finite branch length.}
}
\value{
A one-tip `phylo` tree.
}
\description{
Construct a one-tip tree
}
