% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{get_species_name}
\alias{get_species_name}
\title{Parse species names from labels}
\usage{
get_species_name(a, species_parser = "legacy", sep = "_")
}
\arguments{
\item{a}{One or more labels.}

\item{species_parser}{Species-label convention.}

\item{sep}{Literal input separator.}
}
\value{
Species names separated by spaces.
}
\description{
Parse species names from labels
}
\examples{
get_species_name("Homo_sapiens_gene1")
get_species_name("Genus_cf_species_gene1", species_parser="taxonomic")
}
