% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfspecies.R
\name{get_species_names}
\alias{get_species_names}
\title{Parse species names from tree tips}
\usage{
get_species_names(phy, sep = "_", species_parser = "legacy")
}
\arguments{
\item{phy}{A `phylo` tree.}

\item{sep}{Literal input and output separator.}

\item{species_parser}{Species-label convention.}
}
\value{
A character vector aligned with tree tips.
}
\description{
Parse species names from tree tips
}
