% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfpcm.R
\name{get_tree_table}
\alias{get_tree_table}
\title{Summarize a comparative-model fit}
\usage{
get_tree_table(pcm_out, mode, species_parser = "legacy", sep = "_")
}
\arguments{
\item{pcm_out}{An l1ou or PhylogeneticEM result object.}

\item{mode}{Either `"l1ou"` or `"PhylogeneticEM"`.}

\item{species_parser}{Species-label convention.}

\item{sep}{Literal separator in leaf labels.}
}
\value{
A one-row model summary data frame.
}
\description{
Summarize a comparative-model fit
}
