% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfutil.R
\name{is.blank}
\alias{is.blank}
\title{Test whether a value is blank}
\usage{
is.blank(x, false.triggers = FALSE)
}
\arguments{
\item{x}{An object to inspect.}

\item{false.triggers}{Whether logical false values should count as blank.}
}
\value{
A single logical value.
}
\description{
Test whether a value is blank
}
