% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{is_root}
\alias{is_root}
\title{Test whether nodes are the root}
\usage{
is_root(phy, node_num)
}
\arguments{
\item{phy}{A `phylo` tree.}

\item{node_num}{One or more integer node numbers.}
}
\value{
A logical vector.
}
\description{
Test whether nodes are the root
}
