% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-rooting.R
\name{is_same_root}
\alias{is_same_root}
\title{Compare root splits between trees}
\usage{
is_same_root(phy1, phy2)
}
\arguments{
\item{phy1, phy2}{Rooted `phylo` trees containing the same tips.}
}
\value{
A single logical value.
}
\description{
Compare root splits between trees
}
