% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-imputation.R
\name{phylogenetic_imputation}
\alias{phylogenetic_imputation}
\title{Impute missing traits with Rphylopars}
\usage{
phylogenetic_imputation(tree, trait_table, verbose = FALSE)
}
\arguments{
\item{tree}{A `phylo` tree.}

\item{trait_table}{A numeric trait table named by tree tips.}

\item{verbose}{Whether to report the number of imputed values.}
}
\value{
A data frame of reconstructed tip traits.
}
\description{
Impute missing traits with Rphylopars
}
