% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-transform.R
\name{collapse_short_external_edges}
\alias{collapse_short_external_edges}
\title{Pad short external edges}
\usage{
collapse_short_external_edges(tree, threshold = 1e-06, verbose = FALSE)
}
\arguments{
\item{tree}{A \code{phylo} tree with branch lengths.}

\item{threshold}{Finite non-negative minimum edge length.}

\item{verbose}{Whether to emit progress messages.}
}
\value{
A modified \code{phylo} tree.
}
\description{
Convenience wrapper around \code{\link[=pad_short_edges]{pad_short_edges()}} restricted to terminal edges.
}
