% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-transform.R
\name{contains_polytomy}
\alias{contains_polytomy}
\title{Test whether a tree contains a polytomy}
\usage{
contains_polytomy(phy)
}
\arguments{
\item{phy}{A \code{phylo} tree.}
}
\value{
A single logical value.
}
\description{
Test whether a tree contains a polytomy
}
