% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-traversal.R
\name{get_children_num}
\alias{get_children_num}
\title{Get child node numbers}
\usage{
get_children_num(phy, node_num)
}
\arguments{
\item{phy}{A `phylo` tree.}

\item{node_num}{One or more integer node numbers.}
}
\value{
An integer vector of child node numbers.
}
\description{
Get child node numbers
}
