% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfpcm.R
\name{get_expression_bases}
\alias{get_expression_bases}
\title{Get base names of expression replicate columns}
\usage{
get_expression_bases(trait_table, replicate_sep)
}
\arguments{
\item{trait_table}{A table with non-empty column names.}

\item{replicate_sep}{Literal separator before replicate suffixes.}
}
\value{
Unique base expression names excluding `"gene"`.
}
\description{
Get base names of expression replicate columns
}
