% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tables.R
\name{get_leaf_regimes}
\alias{get_leaf_regimes}
\title{Assign comparative-model regimes to leaves}
\usage{
get_leaf_regimes(pcm_out, mode)
}
\arguments{
\item{pcm_out}{An l1ou or PhylogeneticEM result object.}

\item{mode}{Either `"l1ou"` or `"PhylogeneticEM"`.}
}
\value{
A data frame mapping leaf labels to regimes.
}
\description{
Assign comparative-model regimes to leaves
}
