% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{get_nearest_tips}
\alias{get_nearest_tips}
\title{Find nearest subject tips}
\usage{
get_nearest_tips(phy, query, subjects, mrca_matrix)
}
\arguments{
\item{phy}{A `phylo` tree.}

\item{query}{A single tip label.}

\item{subjects}{Candidate tip labels.}

\item{mrca_matrix}{A named matrix returned by `ape::mrca()`.}
}
\value{
A list containing nearest tips and their MRCA.
}
\description{
Find nearest subject tips
}
