% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{get_node_age}
\alias{get_node_age}
\title{Get the age of a node in an ultrametric tree}
\usage{
get_node_age(phy, node_num)
}
\arguments{
\item{phy}{An ultrametric `phylo` tree.}

\item{node_num}{A single integer node number.}
}
\value{
Numeric time from the node to its descendant tips.
}
\description{
Get the age of a node in an ultrametric tree
}
