% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{get_node_name_by_num}
\alias{get_node_name_by_num}
\title{Look up phylogenetic node labels by number}
\usage{
get_node_name_by_num(phy, node_num)
}
\arguments{
\item{phy}{A `phylo` tree.}

\item{node_num}{One or more integer node numbers. Missing and out-of-range
values are ignored.}
}
\value{
A character vector of node labels.
}
\description{
Look up phylogenetic node labels by number
}
