% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-rooting.R
\name{get_outgroup}
\alias{get_outgroup}
\title{Get the smallest root-child clade}
\usage{
get_outgroup(phy)
}
\arguments{
\item{phy}{A rooted \code{phylo} tree.}
}
\value{
Tip labels in the smallest root-child clade.
}
\description{
Get the smallest root-child clade
}
