% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-traversal.R
\name{get_parent_num}
\alias{get_parent_num}
\title{Get parent node numbers}
\usage{
get_parent_num(phy, node_num)
}
\arguments{
\item{phy}{A \code{phylo} tree.}

\item{node_num}{One or more integer node numbers.}
}
\value{
An integer vector of parent node numbers.
}
\description{
Get parent node numbers
}
