% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-placeholders.R
\name{get_placeholder_leaf}
\alias{get_placeholder_leaf}
\title{Build a placeholder leaf table}
\usage{
get_placeholder_leaf(tree, original_trait_table)
}
\arguments{
\item{tree}{A `phylo` tree.}

\item{original_trait_table}{Original traits named by tree tips.}
}
\value{
A long-format placeholder leaf table.
}
\description{
Build a placeholder leaf table
}
