% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-placeholders.R
\name{get_placeholder_regime}
\alias{get_placeholder_regime}
\title{Build a placeholder regime table}
\usage{
get_placeholder_regime(tree, original_trait_table)
}
\arguments{
\item{tree}{A \code{phylo} tree.}

\item{original_trait_table}{Original numeric trait table.}
}
\value{
A placeholder regime table.
}
\description{
Build a placeholder regime table
}
