% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-placeholders.R
\name{get_placeholder_tree}
\alias{get_placeholder_tree}
\title{Build a placeholder tree-level model summary}
\usage{
get_placeholder_tree(tree, original_trait_table)
}
\arguments{
\item{tree}{A \code{phylo} tree.}

\item{original_trait_table}{Original numeric trait table.}
}
\value{
A one-row placeholder model summary.
}
\description{
Build a placeholder tree-level model summary
}
