% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{get_root_num}
\alias{get_root_num}
\title{Get the root node number}
\usage{
get_root_num(phy)
}
\arguments{
\item{phy}{A `phylo` tree.}
}
\value{
The integer root-node number, or multiple values for malformed trees.
}
\description{
Get the root node number
}
\examples{
tree <- ape::read.tree(text="((A:1,B:1):1,C:1);")
get_root_num(tree)
}
