% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{get_sister_num}
\alias{get_sister_num}
\title{Get sister node numbers}
\usage{
get_sister_num(phy, node_num)
}
\arguments{
\item{phy}{A `phylo` tree.}

\item{node_num}{One or more integer node numbers.}
}
\value{
An integer vector of sibling node numbers.
}
\description{
Get sister node numbers
}
