% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{get_tip_labels}
\alias{get_tip_labels}
\title{Get tip labels below nodes}
\usage{
get_tip_labels(phy, node_num, out = NULL)
}
\arguments{
\item{phy}{A `phylo` tree.}

\item{node_num}{One or more integer node numbers.}

\item{out}{Optional character values to prepend for compatibility.}
}
\value{
A character vector of descendant tip labels.
}
\description{
Get tip labels below nodes
}
