% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-transform.R
\name{has_same_leaves}
\alias{has_same_leaves}
\title{Compare descendant leaves at two nodes}
\usage{
has_same_leaves(phy1, phy1_node, phy2, phy2_node)
}
\arguments{
\item{phy1, phy2}{\code{phylo} trees with identical tip-label multisets.}

\item{phy1_node, phy2_node}{Integer node numbers to compare.}
}
\value{
A single logical value.
}
\description{
Compare descendant leaves at two nodes
}
