% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{is_leaf}
\alias{is_leaf}
\title{Test whether a node is a leaf}
\usage{
is_leaf(phy, node_num)
}
\arguments{
\item{phy}{A `phylo` tree.}

\item{node_num}{A single integer node number.}
}
\value{
A single logical value.
}
\description{
Test whether a node is a leaf
}
