% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-collapse.R
\name{leaf_table_collapse2original}
\alias{leaf_table_collapse2original}
\title{Expand a collapsed-tree leaf table}
\usage{
leaf_table_collapse2original(
  leaf_table,
  tree_original,
  tree_collapsed,
  node_num_mapping
)
}
\arguments{
\item{leaf_table}{A collapsed-tree leaf table.}

\item{tree_original}{Original `phylo` tree.}

\item{tree_collapsed}{Collapsed `phylo` tree.}

\item{node_num_mapping}{Mapping returned by [map_node_num()].}
}
\value{
A leaf table expanded to original tips.
}
\description{
Expand a collapsed-tree leaf table
}
