% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-transform.R
\name{multi2bi_node_number_transfer}
\alias{multi2bi_node_number_transfer}
\title{Map polytomy nodes to a binary resolution}
\usage{
multi2bi_node_number_transfer(
  multifurcated_tree,
  bifurcated_tree,
  verbose = FALSE
)
}
\arguments{
\item{multifurcated_tree}{A multifurcating \code{phylo} tree.}

\item{bifurcated_tree}{A binary resolution with identical ordered tips.}

\item{verbose}{Whether to emit progress messages.}
}
\value{
A data frame mapping multifurcated to binary node numbers.
}
\description{
Map polytomy nodes to a binary resolution
}
