% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfphylo.R
\name{pad_short_edges}
\alias{pad_short_edges}
\title{Pad short phylogenetic edges}
\usage{
pad_short_edges(
  tree,
  threshold = 1e-06,
  external_only = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{tree}{A binary `phylo` tree with branch lengths.}

\item{threshold}{Finite non-negative minimum edge length.}

\item{external_only}{Whether to modify only terminal edges.}

\item{verbose}{Whether to emit progress messages.}
}
\value{
A modified `phylo` tree.
}
\description{
Pad short phylogenetic edges
}
