% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfreconcil.R
\name{read_notung_parsable}
\alias{read_notung_parsable}
\title{Read NOTUNG parsable duplication records}
\usage{
read_notung_parsable(file, mode = "D")
}
\arguments{
\item{file}{Path to a NOTUNG parsable-output file.}

\item{mode}{Event mode. Currently only `"D"` is supported.}
}
\value{
A data frame of duplication-event fields.
}
\description{
Read NOTUNG parsable duplication records
}
