% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-collapse.R
\name{regime_table_collapse2original}
\alias{regime_table_collapse2original}
\title{Restore original node names in a regime table}
\usage{
regime_table_collapse2original(
  regime_table,
  tree_original,
  tree_collapsed,
  node_num_mapping
)
}
\arguments{
\item{regime_table}{A collapsed-tree regime table.}

\item{tree_original}{Original \code{phylo} tree.}

\item{tree_collapsed}{Collapsed \code{phylo} tree.}

\item{node_num_mapping}{Mapping returned by \code{\link[=map_node_num]{map_node_num()}}.}
}
\value{
A regime table using original-tree node names.
}
\description{
Restore original node names in a regime table
}
