% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfpcm.R
\name{regime_table_collapse2original}
\alias{regime_table_collapse2original}
\title{Restore original node names in a regime table}
\usage{
regime_table_collapse2original(
  regime_table,
  tree_original,
  tree_collapsed,
  node_num_mapping
)
}
\arguments{
\item{regime_table}{A collapsed-tree regime table.}

\item{tree_original}{Original `phylo` tree.}

\item{tree_collapsed}{Collapsed `phylo` tree.}

\item{node_num_mapping}{Mapping returned by [map_node_num()].}
}
\value{
A regime table using original-tree node names.
}
\description{
Restore original node names in a regime table
}
