% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfpcm.R
\name{remove_invariant_traits}
\alias{remove_invariant_traits}
\title{Remove invariant trait columns}
\usage{
remove_invariant_traits(trait_table, small_dif = 0.001, verbose = FALSE)
}
\arguments{
\item{trait_table}{A numeric matrix or data frame.}

\item{small_dif}{Finite non-negative range below which a trait is removed.}

\item{verbose}{Whether to report removed traits.}
}
\value{
A list with the filtered table and removed trait names.
}
\description{
Remove invariant trait columns
}
\examples{
remove_invariant_traits(data.frame(a=c(1, 1), b=c(1, 2)))
}
