% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-transform.R
\name{remove_redundant_root_edge}
\alias{remove_redundant_root_edge}
\title{Collapse redundant singleton root edges}
\usage{
remove_redundant_root_edge(phy)
}
\arguments{
\item{phy}{A \code{phylo} tree.}
}
\value{
A \code{phylo} tree with singleton nodes collapsed where applicable.
}
\description{
Collapse redundant singleton root edges
}
