\name{rkftools-functions}
\alias{rkftools-functions}
\title{Function Reference for rkftools}
\description{Index page for functions provided by the \pkg{rkftools} package.}
\details{This page exists to provide documentation entries for package functions. Refer to source code for implementation details.}
\alias{calc_complementarity}
\alias{collapse_clades}
\alias{collapse_short_branches}
\alias{collapse_short_external_edges}
\alias{contains_polytomy}
\alias{count_foreground_lineage}
\alias{fill_node_labels}
\alias{force_ultrametric}
\alias{get_ancestor_num}
\alias{get_bootstrap_table}
\alias{get_children_num}
\alias{get_deepest_node_num}
\alias{get_descendent_num}
\alias{get_duplication_confidence_score}
\alias{get_expression_bases}
\alias{get_high_similarity_clades}
\alias{get_leaf_regimes}
\alias{get_leaf_table}
\alias{get_nearest_tips}
\alias{get_node_age}
\alias{get_node_name_by_num}
\alias{get_node_num_by_name}
\alias{get_outgroup}
\alias{get_parent_num}
\alias{get_parsed_args}
\alias{get_phy2_root_in_phy1}
\alias{get_placeholder_leaf}
\alias{get_placeholder_regime}
\alias{get_placeholder_tree}
\alias{get_regime_table}
\alias{get_root_num}
\alias{get_root_position_dependent_species_overlap_scores}
\alias{get_rooted_newick}
\alias{get_single_branch_tree}
\alias{get_sister_num}
\alias{get_species_name}
\alias{get_species_names}
\alias{get_species_overlap_score}
\alias{get_tip_labels}
\alias{get_tip_labels_obsolete}
\alias{get_tree_table}
\alias{has_same_leaves}
\alias{is_leaf}
\alias{is_root}
\alias{is_same_root}
\alias{is.blank}
\alias{leaf_table_collapse2original}
\alias{leaf2species}
\alias{MAD}
\alias{MAD_parallel}
\alias{map_node_num}
\alias{merge_replicates}
\alias{multi2bi_node_number_transfer}
\alias{pad_short_edges}
\alias{phylogenetic_imputation}
\alias{read_notung_parsable}
\alias{regime_table_collapse2original}
\alias{remove_invariant_traits}
\alias{remove_redundant_root_edge}
\alias{restore_imputed_leaves}
\alias{sort_exp}
\alias{table2phylo}
\alias{phylo2table}
\alias{transfer_node_labels}
\alias{tree_table_collapse2original}

\keyword{utilities}
