% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-labels.R
\name{transfer_node_labels}
\alias{transfer_node_labels}
\title{Transfer internal-node labels between congruent trees}
\usage{
transfer_node_labels(phy_from, phy_to)
}
\arguments{
\item{phy_from}{A labeled \code{phylo} tree.}

\item{phy_to}{A \code{phylo} tree with the same tips.}
}
\value{
\code{phy_to} with labels transferred by clade signature.
}
\description{
Transfer internal-node labels between congruent trees
}
